pemilu19_2/requirements.txt
absl-py==0.6.1
alabaster==0.7.10
anaconda-client==1.6.14
anaconda-navigator==1.8.7
anaconda-project==0.8.2
asn1crypto==0.24.0
astor==0.7.1
astroid==1.6.3
astropy==3.0.2
attrs==18.1.0
aylien-apiclient==0.7.0
Babel==2.5.3
backcall==0.1.0
backports.csv==1.0.6
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.3
bokeh==0.12.16
boto==2.48.0
boto3==1.9.61
botocore==1.12.61
Bottleneck==1.2.1
bz2file==0.98
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cheroot==6.5.2
CherryPy==18.1.0
click==6.7
cloudpickle==0.5.3
clyent==1.2.2
colorama==0.3.9
comtypes==1.1.4
conda==4.5.11
conda-build==3.10.5
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.2.2
cycler==0.10.0
cymem==1.31.2
Cython==0.28.2
cytoolz==0.9.0.1
dask==0.17.5
datashape==0.5.4
decorator==4.3.0
dill==0.2.8.2
distributed==1.21.8
-e git+https://github.com/django/django.git@b709d701303b3877387020c1558a590713b09853#egg=Django
django-pandas==0.6.0
docutils==0.14
en-core-web-sm==2.0.0
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
feedparser==5.2.1
filelock==3.0.4
Flask==1.0.2
Flask-Cors==3.0.4
funcy==1.11
future==0.17.1
gast==0.2.0
gensim==3.6.0
gevent==1.3.0
glob2==0.6
greenlet==0.4.13
grpcio==1.17.1
h5py==2.7.1
heapdict==1.0.0
helper==2.4.2
html5lib==1.0.1
httplib2==0.12.0
idna==2.6
imageio==2.3.0
imagesize==1.0.0
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==0.24
jaraco.functools==1.20
jdcal==1.4
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
joblib==0.13.0
jsonpickle==1.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.23.1
locket==0.2.0
lxml==4.2.1
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.3
mkl-fft==1.0.0
mkl-random==1.0.1
more-itertools==4.1.0
mpmath==1.0.0
msgpack-numpy==0.4.3.1
msgpack-python==0.5.6
multipledispatch==0.5.0
murmurhash==0.28.0
mysqlclient==1.3.14
navigator-updater==0.2.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==5.5.0
numba==0.38.0
numexpr==2.6.5
numpy==1.14.3
numpydoc==0.8.0
oauthlib==2.1.0
odo==0.5.1
olefile==0.45.1
openpyxl==2.5.3
packaging==17.1
pandas==0.23.0
pandocfilters==1.4.2
parso==0.2.0
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
patsy==0.5.0
Pattern==3.6
pdfminer.six==20181108
pep8==1.7.1
pickleshare==0.7.4
Pillow==5.1.0
pkginfo==1.4.2
plac==0.9.6
plotly==3.4.2
pluggy==0.6.0
ply==3.11
portend==2.3
preshed==1.0.0
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==5.4.5
py==1.5.3
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.7.2
pycurl==7.43.0.1
pyflakes==1.6.0
Pygments==2.2.0
pyLDAvis==2.1.2
pylint==1.8.4
PyMySQL==0.9.3
pyodbc==4.0.23
pyOpenSSL==18.0.0
pyparsing==2.2.0
pyreadline==2.1
PySocks==1.6.8
pytest==3.5.1
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
python-dateutil==2.7.3
python-docx==0.8.7
pytz==2018.4
PyWavelets==0.5.2
pywin32==223
pywinpty==0.5.1
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.1
regex==2017.11.9
requests==2.18.4
requests-oauthlib==1.0.0
retrying==1.3.3
reverse-geocoder==1.5.1
rope==0.10.7
ruamel-yaml==0.15.35
s3transfer==0.1.13
Sastrawi==1.0.1
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
smart-open==1.7.1
snowballstemmer==1.2.1
sortedcollections==0.6.1
sortedcontainers==1.5.10
spacy==2.0.12
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
SQLAlchemy==1.2.7
sqlparse==0.2.4
statsmodels==0.9.0
stop-words==2018.7.23
sympy==1.1.1
tables==3.4.3
tblib==1.3.2
tempora==1.14
tensorboard==1.12.1
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
textblob==0.15.2
thinc==6.10.3
toolz==0.9.0
tornado==5.0.2
tqdm==4.26.0
traitlets==4.3.2
tweepy==3.7.0
typing==3.6.4
ujson==1.35
unicode==2.6
unicodecsv==0.14.1
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wordcloud==1.5.0
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.4
xlwings==0.11.8
xlwt==1.3.0
zc.lockfile==1.4
zict==0.1.3
